:arrow_up: Migrate from legacy Endpoints to EndpointSlice
https://kubernetes.io/docs/concepts/services-networking/service/#what-s-next
marks the "Endpoints" as legacy.

The [EndpointSlices] documentation mentiones that it is an alternative
to Endpoints.

On my cluster the Endpoints resource let Kubernetes spawn an
EndpointsSlice resources with the same settings.

So I assume the Endpoints resource might be deprecated somewhere in the
future.

Tested with similar configuration in on my personal cluster:
https://github.com/rappet/stash/commit/b547479018d7859c5521b3e489ad3334bbdb923d

[EndpointSlices]: https://kubernetes.io/docs/concepts/services-networking/endpoint-slices/